+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
-All Notable changes to `menu` will be documented in this file
+All Notable changes to `menu` will be documented in this file.
 
 ## 1.2.0
 - Broadened the api of `Menu`:
     - A `submenu` method
     - A `blueprint` method to copy the menu without it's contents
-    - More convenience methods: `addLinkClass`, `addLinkParentClass`, `addLinkParentAttribute`, `addItemParentClass`, `setItemParentAttribute`
+    - Html item convenience methods: `addItemClass`, `setItemAttribute`
+    - Html parent convenience methods: `addItemParentClass`, `setItemParentAttribute`
 - Added `HasHtmlAttributes` and `HasParentAttributes` interfaces
 
 ## 1.1.1