+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 All Notable changes to `menu` will be documented in this file.
 
 ## 1.2.0
-- Broadened the api of `Menu`:
-    - `submenu`
+- New methods on `Menu`:
+    - `submenu` for submenus with optional headers
     - `void` and `voidIf` for empty list items
+    - `wrap` to wrap the menu in an html tag with optional attributes
     - A `blueprint` method to copy the menu without it's contents
     - Html item convenience methods: `addItemClass`, `setItemAttribute`
     - Html parent convenience methods: `addItemParentClass`, `setItemParentAttribute`